fix error & completion method
