+++ src/service/helpers/index.ts
@@ -1,3 +1,3 @@
 export * from './map-to-if-defined'
 export * from './mutation-handlers'
-export * from './notify-not-found'
+export * from './notifications'
