--- src/service/helpers/notify-not-found.ts
@@ -1,8 +1,0 @@
-import { NotFoundHandler, MutationType } from '../../types'
-
-export const notifyNotFound = (
-  type: MutationType,
-  notFoundHandler?: NotFoundHandler,
-  isAll?: boolean
-) =>
-  notFoundHandler ? notFoundHandler(type, isAll) : console.log('DATA NOT FOUND')
